Polyu course COMP1002 Assignment 1 Problem Two
Write a Python program that accepts an integer from the keyboard. Assume the input by the user is always positive and 4-digit long. Based on your idea in Q1, add all digits in the integer and print the sum out to the screen. e.g. if the number is 1974, your program will output 21 since 1+9+7+4=21.